Exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -17,9 +17,12 @@ blogsRouter.post('', async (request, response, next) => {
   try {
     logger.info(request.body)
     const body = request.body
+    if (!request.token) {
+      return response.status(401).json({ error: "token missing" })
+    }
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!request.token || !decodedToken.id) {
-      return response.status(401).json({ error: 'token missing or invalid' })
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
     }
     const user = await User.findById(decodedToken.id)
 
@@ -41,8 +44,22 @@ blogsRouter.post('', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    if (!request.token) {
+      return response.status(401).json({ error: "token missing" })
+    }
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    if (blog.user.toString() === user._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: "blog can only be deleted by user who added it" })
+    }
   } catch (exception) {
     next(exception)
   }
